Fix bug in ParticipantInteraction logic

+++ src/main/kotlin/com/avilanii/backttend/presentation/routes/ParticipantsRoutes.kt
@@ -1,7 +1,6 @@
 package com.avilanii.backttend.presentation.routes
 
 import com.avilanii.backttend.domain.models.Participant
-import com.avilanii.backttend.domain.models.ParticipantInteraction
 import com.avilanii.backttend.infrastructure.datatransferobjects.CheckInConfirmationDTO
 import com.avilanii.backttend.infrastructure.datatransferobjects.toParticipantResponseDTO
 import com.avilanii.backttend.services.ParticipantService
@@ -28,7 +27,7 @@ fun Route.participantsRoutes(
             post<Participant> { participant ->
                 val eventId = call.parameters["id"]?.toIntOrNull() ?: return@post call.respondText("No id provided", status = HttpStatusCode.NotFound)
                 val userId = userService.findByEmail(participant.email)?.id
-                val participantId = participantService.addParticipant(participant.copy(eventId = eventId.toInt(), userId = userId, qrCode = UUID.randomUUID().toString()))
+                val participantId = participantService.addParticipant(participant.copy(eventId = eventId, userId = userId, qrCode = UUID.randomUUID().toString()))
                 val addedParticipant = participantService.getParticipantById(participantId) ?: return@post call.respondText("Participant not created", status = HttpStatusCode.BadRequest)
                 call.respond(HttpStatusCode.OK, addedParticipant)
             }
